Fix a build break on 64 bit (uint and size_t are distinct types).

+++ sql/sql_lex.h
@@ -1782,7 +1782,7 @@ typedef struct st_lex : public Query_tables_list
     context_stack.pop();
   }
 
-  bool copy_db_to(char **p_db, uint *p_db_length) const;
+  bool copy_db_to(char **p_db, size_t *p_db_length) const;
 
   Name_resolution_context *current_context()
   {
